--- src/renderer/components/atoms/dropdown/index.ts
@@ -1,1 +1,0 @@
-export { default } from './dropdown';
