+++ src/renderer/components/atoms/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
 export { default as Text } from './text';
 export { default as Input } from './input';
-export { default as Dropdown } from './dropdown';
+export { default as SelectionList } from './selection-list';
 export * from './button';
