+++ src/renderer/components/atoms/selection-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './selection-list';
